product display has been improved
