+++ src/CompanyLogo/CompanyLogo.tsx
@@ -1,11 +1,10 @@
 import React from "react";
-
+const imageUrl = new URL("../../src/assets/Logo_right.svg", import.meta.url).href;
 const CompanyLogo: React.FC = () => {
   return (
     <div>
-      <img src="../../src/assets/Logo_right.svg" alt="" width={70} height={50} />
+      <img src={imageUrl} alt="" width={70} height={50} />
     </div>
   );
 };
-
 export default CompanyLogo;
