+++ src/components/Products/Product.styled.ts
@@ -28,24 +28,24 @@ export const Arrow = styled.button`
 export const ProductContainer = styled.div`
   position: static;
   display: flex;
-  flex-wrap: wrap;
-  justify-content: space-evenly;
-  align-content: space-around;
+  /* flex-wrap: wrap; */
+  /* justify-content: space-evenly; */
+  /* align-content: space-around; */
   align-items: center;
   height: 50vh;
-  width: 92vw;
+  width: 100vw;
   border-radius: 10px;
   background-color: var(--color-very-light-grey);
   transition: 0.2s linear;
   border: none;
-  gap: 10px;
+  /* gap: 10px; */
 `;
 
 export const Container = styled.button`
   position: static;
   display: flex;
-  justify-content: center;
-  align-items: flex-start;
+  /* justify-content: center; */
+  /* align-items: flex-start; */
   align-content: space-around;
   flex-wrap: wrap;
   font-size: 1rem;
