+++ src/pages/MainPage.tsx
@@ -1,5 +1,6 @@
 import BottomNavigation from "../components/BottomNavigation/BottomNavigation.component";
-import { Products } from "../components/Products/Product.component";
+
+import { Products } from "../components/Product/Product.component";
 
 export const MainPage = () => {
   return (
